Restructure devtools tests
