+++ debug/test/browser/devtools/util.test.js
@@ -1,8 +1,8 @@
-import { setupScratch, teardown } from '../../../test/_util/helpers';
+import { setupScratch, teardown } from '../../../../test/_util/helpers';
 import { h, Component, Fragment } from 'preact';
 import { forwardRef, memo } from 'preact/compat';
-import { setIn } from '../../src/devtools/util';
-import { getDisplayName } from '../../src/devtools/vnode';
+import { setIn } from '../../../src/devtools/util';
+import { getDisplayName } from '../../../src/devtools/vnode';
 
 /** @jsx h */
